The engine previously overvalued attacking advantages and threatened pieces, leading it to make unsound material sacrifices (like N+B for R+P). This was caused by: 1. A bug in the king zone calculation that made it asymmetric for Black. 2. Extremely aggressive King Safety and Threat constants. 3. Traditional material values that undervalued minor pieces in the middlegame.
I fixed the bug, tuned down the aggressive evaluation components, and adjusted material values to discourage these exchanges. Verification shows the reported blunder is now avoided.

+++ chess_engine/see.py
@@ -93,7 +93,7 @@ def get_lva_and_remove(attackers, piece_bbs, side_mask):
         subset = attackers & piece_bbs[i + offset]
         if subset:
             sq_bb = subset & -subset
-            return PIECE_VALUES[i], sq_bb, i
+            return PIECE_VALUES[i], np.uint64(sq_bb), i
     return 0, np.uint64(0), -1
 
 @numba.njit(cache=True)
